# services/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Gamics API
  version: 0.9.0
  description: >-
    Gamics player, competition, blind match result verification, Gamics result review, rankings, M-Pesa payments,
    notification, identity and role-based operations API.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/auth/register:
    post:
      operationId: registerPlayer
      tags:
        - Authentication
      summary: Register with a username, eFootball Konami ID and password.
      description: >-
        Registration asks for nothing else. The username becomes the player's handle and display name, the Konami ID
        becomes their eFootball game account, and the current terms and privacy notice, which the registration
        screen must present, are recorded as accepted. The new player can enter competitions straight away. Email
        and phone are added later from the signed-in account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - username
                - konamiId
                - password
              properties:
                username:
                  type: string
                  pattern: ^[A-Za-z0-9][A-Za-z0-9_.]{2,23}$
                konamiId:
                  type: string
                  pattern: ^[A-Za-z0-9][A-Za-z0-9 _.-]{5,63}$
                  description: >-
                    As shown in eFootball. Case, spaces, dots, dashes and underscores are ignored when comparing, and
                    at least six letters or digits are required.
                password:
                  type: string
                  minLength: 8
                  maxLength: 128
                  description: Characters, not bytes. It may not be the username or the Konami ID.
                deviceName:
                  type: string
                  maxLength: 120
      responses:
        '201':
          description: Registered and signed in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '400':
          $ref: '#/components/responses/Error'
          description: invalid_username, invalid_konami_id, invalid_password, weak_password, invalid_device_name.
        '409':
          $ref: '#/components/responses/Error'
          description: username_taken; konami_id_taken (the Konami ID already has an account, so sign in instead).
        '429':
          $ref: '#/components/responses/Error'
          description: rate_limited (too many registrations from this network), with Retry-After.
        '503':
          $ref: '#/components/responses/Error'
  /v1/auth/login:
    post:
      operationId: signIn
      tags:
        - Authentication
      summary: Sign in with a Konami ID and password.
      description: >-
        An unknown Konami ID and a wrong password return the same 401. Repeated failures for one Konami ID or one
        network are refused with 429 for LOGIN_FAILURE_WINDOW.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - konamiId
                - password
              properties:
                konamiId:
                  type: string
                password:
                  type: string
                deviceName:
                  type: string
                  maxLength: 120
      responses:
        '200':
          description: Authenticated session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '401':
          $ref: '#/components/responses/Error'
          description: invalid_credentials.
        '403':
          $ref: '#/components/responses/Error'
          description: account_unavailable (suspended), only after a correct password.
        '429':
          $ref: '#/components/responses/Error'
          description: rate_limited, with Retry-After.
        '503':
          $ref: '#/components/responses/Error'
  /v1/auth/password-reset/request:
    post:
      operationId: requestPasswordReset
      tags:
        - Authentication
      summary: Email a password-reset code to the account's verified address.
      description: >-
        Always 202 for a well-formed Konami ID, whether or not it has an account or a verified email, so the answer
        reveals nothing. Players who never verified an email cannot reset their password themselves; the app should
        ask players to add one.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - konamiId
              properties:
                konamiId:
                  type: string
      responses:
        '202':
          description: Accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodeAccepted'
        '400':
          $ref: '#/components/responses/Error'
          description: invalid_konami_id.
        '429':
          $ref: '#/components/responses/Error'
          description: rate_limited, with Retry-After.
        '503':
          $ref: '#/components/responses/Error'
  /v1/auth/password-reset/confirm:
    post:
      operationId: confirmPasswordReset
      tags:
        - Authentication
      summary: Set a new password with the emailed code and sign every device out.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - konamiId
                - code
                - newPassword
              properties:
                konamiId:
                  type: string
                code:
                  type: string
                  pattern: ^[0-9]{6}$
                newPassword:
                  type: string
                  minLength: 8
                  maxLength: 128
      responses:
        '204':
          description: Password changed. Sign in with the new password.
        '400':
          $ref: '#/components/responses/Error'
          description: invalid_password, weak_password. The code is not spent.
        '401':
          $ref: '#/components/responses/Error'
          description: invalid_code (wrong, expired or used).
        '503':
          $ref: '#/components/responses/Error'
  /v1/auth/refresh:
    post:
      operationId: refreshSession
      tags:
        - Authentication
      summary: Rotate a refresh token and issue a new access token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Rotated authenticated session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '401':
          $ref: '#/components/responses/Error'
  /v1/auth/logout:
    post:
      operationId: logout
      tags:
        - Authentication
      description: >-
        Revokes the current session and every push installation registered with it, so the device stops receiving
        the player's pushes. The app should still call DELETE /v1/me/push-tokens/{id} for its installation first:
        an installation registered with an earlier session of the same device is not bound to this one.
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Current refresh session and its push installations revoked.
        '401':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /v1/me:
    get:
      operationId: getCurrentPlayer
      tags:
        - Players
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current player.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        '401':
          $ref: '#/components/responses/Error'
    patch:
      operationId: updateCurrentPlayer
      tags:
        - Players
      description: Optional profile details a player can change after registration.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              minProperties: 1
              properties:
                displayName:
                  type: string
                  minLength: 2
                  maxLength: 80
                  description: The name opponents and public profiles show. It starts as the username.
                countryCode:
                  type: string
                  pattern: ^[A-Z]{2}$
                birthDate:
                  type: string
                  format: date
      responses:
        '200':
          description: Updated player.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        '400':
          $ref: '#/components/responses/Error'
          description: >-
            invalid_display_name (2-80 characters), invalid_country, invalid_birth_date, empty_update,
            versioned_legal_acceptance_required.
  /v1/me/profile:
    put:
      operationId: putPlayerProfile
      tags:
        - Players
      description: The handle is the username. Changing it does not change the display name or the Konami ID.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileInput'
      responses:
        '200':
          description: Updated player and profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        '409':
          $ref: '#/components/responses/Error'
  /v1/me/password:
    post:
      operationId: changePassword
      tags:
        - Players
      summary: Change the password and sign every other device out.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - currentPassword
                - newPassword
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
                  minLength: 8
                  maxLength: 128
      responses:
        '204':
          description: Password changed; this session stays signed in.
        '400':
          $ref: '#/components/responses/Error'
          description: invalid_password, weak_password.
        '401':
          $ref: '#/components/responses/Error'
          description: invalid_credentials (wrong current password), authentication_required, invalid_session.
        '429':
          $ref: '#/components/responses/Error'
          description: rate_limited, with Retry-After.
  /v1/me/email:
    post:
      operationId: requestEmailVerification
      tags:
        - Players
      summary: Add or change the account email; a code is sent to confirm it.
      description: >-
        The address is stored only after POST /v1/me/email/verify. A verified email receives Gamics communication
        and is the only way to reset a forgotten password.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '202':
          description: Code sent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodeAccepted'
        '400':
          $ref: '#/components/responses/Error'
          description: invalid_email.
        '409':
          $ref: '#/components/responses/Error'
          description: email_already_verified, email_taken.
        '429':
          $ref: '#/components/responses/Error'
          description: rate_limited, with Retry-After.
        '503':
          $ref: '#/components/responses/Error'
          description: email_unavailable, database_unavailable.
  /v1/me/email/verify:
    post:
      operationId: verifyEmail
      tags:
        - Players
      summary: Confirm the emailed code and store the address as verified.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - code
              properties:
                code:
                  type: string
                  pattern: ^[0-9]{6}$
      responses:
        '200':
          description: Updated player.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        '401':
          $ref: '#/components/responses/Error'
          description: invalid_code.
        '409':
          $ref: '#/components/responses/Error'
          description: email_taken.
  /v1/me/phone:
    put:
      operationId: putPhone
      tags:
        - Players
      summary: Save the phone number used for payments.
      description: >-
        International E.164 format with the country code, for example +254712345678 or +919812345678. One phone
        number belongs to one account. The number is not verified by SMS. M-Pesa uses it when a payment request
        omits phoneNumber, and M-Pesa itself accepts only Kenyan Safaricom numbers.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - phoneNumber
              properties:
                phoneNumber:
                  type: string
      responses:
        '200':
          description: Updated player.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        '400':
          $ref: '#/components/responses/Error'
          description: invalid_phone (missing country code or not a phone number).
        '409':
          $ref: '#/components/responses/Error'
          description: phone_taken (linked to another account; Gamics support can help the owner).
    delete:
      operationId: deletePhone
      tags:
        - Players
      summary: Remove the saved phone number.
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Removed.
  /v1/me/game-accounts:
    get:
      operationId: listMyGameAccounts
      tags:
        - Game accounts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Connected game accounts.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/GameAccount'
    post:
      operationId: createGameAccount
      tags:
        - Game accounts
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameAccountInput'
      responses:
        '201':
          description: Connected game account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameAccount'
  /v1/me/game-accounts/{id}:
    patch:
      operationId: updateGameAccount
      tags:
        - Game accounts
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                platform:
                  type: string
                  enum:
                    - android
                    - ios
                inGameName:
                  type: string
                  minLength: 2
                  maxLength: 80
                publisherPlayerId:
                  type:
                    - string
                    - 'null'
      responses:
        '200':
          description: Updated game account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameAccount'
        '404':
          $ref: '#/components/responses/Error'
...
  /v1/payments/{id}:
    get:
      operationId: getPayment
      tags:
        - Payments
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: >-
            Current payment status, the registration outcome (registrationStatus) and the latest refund. Only the
            payer can read a payment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentIntent'
        '400':
          $ref: '#/components/responses/Error'
          description: invalid_payment_id (not a UUID).
        '401':
          $ref: '#/components/responses/Error'
          description: authentication_required, invalid_access_token, invalid_session.
        '404':
          $ref: '#/components/responses/Error'
          description: payment_not_found (unknown, or another player's payment).
        '503':
          $ref: '#/components/responses/Error'
          description: database_unavailable, session_check_unavailable.
  /v1/payments/mpesa/callback/{token}:
    post:
      operationId: receiveMPesaCallback
      tags:
        - Provider callbacks
      summary: Daraja-only callback endpoint; mobile clients must never call it.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            minLength: 32
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Callback durably accepted and processed or queued for replay.
        '400':
          $ref: '#/components/responses/Error'
        '404':
          description: Invalid callback token.
        '503':
          $ref: '#/components/responses/Error'
  /v1/me/matches:
    get:
      operationId: listMyMatches
      tags:
        - Matches
      summary: List the signed-in player's active or historical match assignments.
      description: >-
        Uses signed keyset cursors. The server derives lifecycle and allowedActions for the authenticated
        participant; summaries never carry a score claim.
      security:
        - bearerAuth: []
      parameters:
        - name: state
          in: query
          schema:
            type: string
            enum:
              - active
              - history
            default: active
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - name: cursor
          in: query
          schema:
            type: string
            maxLength: 1024
      responses:
        '200':
          description: A bounded page of participant-authorized matches.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchPage'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
...
  /v1/me/registrations:
    get:
      operationId: listMyRegistrations
      tags:
        - Registrations
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - name: cursor
          in: query
          schema:
            type: string
            maxLength: 2048
      responses:
        '200':
          description: Signed-in player's registrations.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationPage'
        '401':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Request failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
...
    AuthSession:
      type: object
      required:
        - accessToken
        - refreshToken
        - tokenType
        - expiresInSeconds
        - player
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          const: Bearer
        expiresInSeconds:
          type: integer
        player:
          $ref: '#/components/schemas/Player'
...
    CodeAccepted:
      type: object
      required:
        - status
        - expiresInSeconds
      properties:
        status:
          type: string
          const: accepted
        expiresInSeconds:
          type: integer
    ProfileInput:
      type: object
      required:
        - handle
        - bio
        - discoverable
        - analyticsConsent
        - scoutingConsent
      properties:
        handle:
          type: string
          pattern: ^[A-Za-z0-9][A-Za-z0-9_.]{2,23}$
        bio:
          type: string
          maxLength: 280
        discoverable:
          type: boolean
        analyticsConsent:
          type: boolean
        scoutingConsent:
          type: boolean
...